refactor(dca/step2): refactor step2 into multiple components
